add dynamic route to board
